requirements.txt
pyannote.audio
torch
torchaudio
transformers
librosa
noisereduce
pydub